+++ ipatests/ipa-test-task
@@ -27,7 +27,7 @@ import argparse
 
 from ipapython.ipa_log_manager import log_mgr, standard_logging_setup
 from ipatests.pytest_plugins.integration import config
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 from ipatests.test_integration.host import Host
 from ipatests.pytest_plugins.integration import collect_logs
