+++ ipatests/test_integration/test_dns_locations.py
@@ -8,7 +8,7 @@ import dns.rdatatype
 import dns.rdataclass
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 from ipapython.dnsutil import DNSName
 
 IPA_DEFAULT_MASTER_SRV_REC = (
