+++ ipatests/test_integration/test_idviews.py
@@ -5,7 +5,7 @@
 import os
 import re
 import string
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 from ipatests.test_integration.base import IntegrationTest
 from ipatests.pytest_plugins.integration.env_config import get_global_config
 from ipaplatform.paths import paths
