+++ ipatests/test_integration/test_replica_promotion.py
@@ -5,12 +5,12 @@
 import time
 import pytest
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
-from ipatests.test_integration.tasks import assert_error
+from ipatests.pytest_plugins.integration import tasks
+from ipatests.pytest_plugins.integration.tasks import (
+    assert_error, replicas_cleanup)
 from ipalib.constants import DOMAIN_LEVEL_0
 from ipalib.constants import DOMAIN_LEVEL_1
 from ipalib.constants import DOMAIN_SUFFIX_NAME
-from ipatests.test_integration.tasks import replicas_cleanup
 
 
 class ReplicaPromotionBase(IntegrationTest):
