+++ ipatests/test_integration/test_replication_layouts.py
@@ -7,7 +7,7 @@ import pytest
 from ipalib.constants import DOMAIN_LEVEL_0
 from ipatests.pytest_plugins.integration.env_config import get_global_config
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 
 config = get_global_config()
