+++ ipatests/test_integration/test_simple_replication.py
@@ -23,7 +23,7 @@ import pytest
 
 from ipapython.dn import DN
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 
 
 @pytest.mark.ds_acceptance
