+++ ipatests/test_integration/test_topologies.py
@@ -17,7 +17,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 
 
 def test_topology_star():
