+++ ipatests/test_integration/test_topology.py
@@ -7,7 +7,7 @@ import re
 import pytest
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 from ipatests.pytest_plugins.integration.env_config import get_global_config
 from ipalib.constants import DOMAIN_SUFFIX_NAME
 from ipatests.util import assert_deepequal
