+++ ipatests/test_integration/test_trust.py
@@ -21,7 +21,7 @@ import nose
 import re
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 from ipatests.test_integration import util
 from ipaplatform.paths import paths
