+++ ipatests/test_integration/test_vault.py
@@ -5,7 +5,7 @@
 import time
 
 from ipatests.test_integration.base import IntegrationTest
-from ipatests.test_integration import tasks
+from ipatests.pytest_plugins.integration import tasks
 
 WAIT_AFTER_ARCHIVE = 30  # give some time to replication
